Re-render Adder when `annotationsForSelection` property changes
This fixes an issue where the `show` method could position the adder
incorrectly if the emptiness of `annotationsForSelection` changed since
the adder was previously displayed. This is because the size of the
adder was outdated at the point when `show` calls `_calculateTarget` to
determine the adder position. Re-rendering as soon as
`annotationsForSelection` changes resolves this.

Fixes #3293